refactoring added data analyse
